Rename NodePopulateStandardAttributes to NodeDoPopulateStandardAttributes
